WIP timing refactoring
