+++ src/Act/Bounds.hs
@@ -7,9 +7,13 @@ import Data.Maybe
 import Data.List (nub)
 
 import Act.Syntax
-import Act.Syntax.Annotated
+import Act.Syntax.Typed
 import Act.Type (globalEnv)
 import Act.Syntax.Timing
+import Act.Syntax.Untyped (makeIface, Pn, Id, SlotType(..), Decl(..), Interface(..), Pointer(..), EthEnv(..), ValueType(..))
+import Act.Syntax.Types
+import Data.Typeable (Typeable)
+import Act.Parse (nowhere)
 
 
 {-|
@@ -50,7 +54,7 @@ addBoundsBehaviour behv@(Behaviour _ _ (Interface _ decls) _ pre cases post stat
       pre' = pre
              <> mkCallDataBounds decls
              <> mkStorageBounds stateUpdates Pre
-             <> mkStorageBoundsLoc (nub $ concatMap locsFromExp (pre <> cases) <> concatMap locsFromUpdateRHS stateUpdates)
+             <> mkStorageBoundsLoc Pre (nub $ concatMap locsFromExp (pre <> cases) <> concatMap locsFromUpdateRHS stateUpdates)
              <> mkEthEnvBounds (ethEnvFromBehaviour behv)
       post' = post
               <> mkStorageBounds stateUpdates Post
@@ -66,10 +70,10 @@ addBoundsInvariant (Constructor _ (Interface _ decls) _ _ _ _ _) inv@(Invariant 
       storagebounds' = storagebounds
                        <> mkStorageBoundsLoc (locsFromExp predicate)
 
-mkEthEnvBounds :: [EthEnv] -> [Exp ABoolean]
+mkEthEnvBounds :: Typeable t => [EthEnv] -> [Exp ABoolean t]
 mkEthEnvBounds vars = catMaybes $ mkBound <$> nub vars
   where
-    mkBound :: EthEnv -> Maybe (Exp ABoolean)
+    mkBound :: EthEnv -> Maybe (Exp ABoolean t)
     mkBound e = case lookup e globalEnv of
       Just AInteger -> Just $ bound (toAbiType e) (IntEnv nowhere e)
       _ -> Nothing
@@ -91,25 +95,25 @@ mkEthEnvBounds vars = catMaybes $ mkBound <$> nub vars
       Nonce -> AbiUIntType 256
 
 -- | Extracts bounds from the AbiTypes of Integer variables in storage
-mkStorageBounds :: [StorageUpdate] -> When -> [Exp ABoolean]
-mkStorageBounds refs whn = concatMap mkBound refs
+mkStorageBounds :: [StorageUpdate] -> When -> [Exp ABoolean Timed]
+mkStorageBounds refs t = concatMap mkBound refs
   where
-    mkBound :: StorageUpdate -> [Exp ABoolean]
-    mkBound (Update SInteger item _) = [mkItemBounds (setTime whn item)]
+    mkBound :: StorageUpdate -> [Exp ABoolean t]
+    mkBound (Update SInteger item _) = [mkItemBounds t item]
     mkBound _ = []
 
-mkItemBounds :: TItem AInteger Storage -> Exp ABoolean
-mkItemBounds item@(Item _ (PrimitiveType vt) _) = bound vt (TEntry nowhere SStorage item)
-mkItemBounds (Item _ (ContractType _) _) = LitBool nowhere True
+mkItemBounds :: Time t -> TItem AInteger Storage t -> Exp ABoolean t
+mkItemBounds whn item@(Item _ (PrimitiveType vt) _) = bound vt (SVarRef nowhere whn item)
+mkItemBounds _ (Item _ (ContractType _) _) = LitBool nowhere True
 
-mkStorageBoundsLoc :: [StorageLocation] -> [Exp ABoolean]
-mkStorageBoundsLoc refs = concatMap mkBound refs
+mkStorageBoundsLoc :: Time t -> [StorageLocation] -> [Exp ABoolean t]
+mkStorageBoundsLoc whn refs = concatMap mkBound refs
   where
-    mkBound :: StorageLocation -> [Exp ABoolean]
-    mkBound (Loc SInteger item) = [mkItemBounds item]
+    mkBound :: StorageLocation -> [Exp ABoolean t]
+    mkBound (Loc SInteger _ item) = [mkItemBounds whn item]
     mkBound _ = []
 
-mkCallDataBounds :: [Decl] -> [Exp ABoolean]
+mkCallDataBounds :: [Decl] -> [Exp ABoolean t]
 mkCallDataBounds = concatMap $ \(Decl typ name) -> case fromAbiType typ of
   AInteger -> [bound typ (_Var typ name)]
   _ -> []
