+++ src/Act/HEVM_utils.hs
@@ -20,8 +20,8 @@ import Control.Monad.ST (stToIO, ST)
 import Control.Monad.Reader
 import Control.Monad 
 
-import Act.Syntax.Annotated
-import Act.Syntax.Untyped (makeIface)
+import Act.Syntax.Typed
+import Act.Syntax.Untyped (Interface(..), Decl(..), makeIface)
 
 import qualified EVM.Types as EVM
 import EVM.Types (VM(..))
